+++ src/fetch/fetchers/characters.js
@@ -1,8 +1,7 @@
-const q = require('q');
+import q from 'q';
+import gw2 from 'lib/gw2';
 
-const gw2 = require('lib/gw2');
-
-module.exports = function fetchUserCharacterData (models, { token }) {
+export default function fetchUserCharacterData (models, { token }) {
   return gw2.readCharactersDeep(token)
     .then((characters) => {
       return models.Gw2Character.destroy({
@@ -42,4 +41,4 @@ module.exports = function fetchUserCharacterData (models, { token }) {
         return q.allSettled([...upsertCharactersPromises]);
       });
     });
-};
+}