+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 - Changes
 ```
 
+## [1.1.0] - 2017/01/07
+### Added
+- PvpStandings table to db
+- Pvp standings data fetcher
+- Current pvp season leaderboard
+
 ## [1.0.0] - 2017/01/02
 ### Added
 - This CHANGELOG file to allow more insight to the changes made throughout the development of api.gw2armory.com